// src/data-type.ts
export type INT8 = number;
export type INT16 = number;
export type INT32 = number;
export type INT64 = number;

export type UINT8 = number;
export type UINT16 = number;
export type UINT32 = number;
export type UINT64 = number;

export enum SIZE {

	INT8 = 1,
	INT16 = 2,
	INT32 = 4,
	INT64 = 8,

	UINT8 = 1,
	UINT16 = 2,
	UINT32 = 4,
	UINT64 = 8,

}